release: 2.0.1
The CodinGame API changed and it's now more difficult to communicate
with the website directly. As such, sendBotCode and getCodinGamerId
commands have been removed, as well as their related settings.

Fixed a bug in the new bot creation code to avoid trying to create a bot
with a name already in use.